requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
crewai==0.11.0
langchain-openai==0.0.2
langchain==0.0.350
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slack-sdk==3.26.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
redis==5.0.1
celery==5.3.4
httpx==0.24.1
aiofiles==23.2.1
